Move calibrator into libs, create example and add orientation checks
