+++ pitop/pma/common/__init__.py
@@ -1,1 +1,2 @@
 from .utils import get_pin_for_port, type_check  # noqa: F401
+from .imu_calibration import ImuCalibration
